+++ Demo/Demo/PayPalVault/PayPalVaultViewModel/PayPalVaultViewModel.swift
@@ -34,7 +34,8 @@ class PayPalVaultViewModel: VaultViewModel {
         let setupToken = try await fetchSetupToken(
             customerID: customerID.isEmpty ? nil : customerID,
             selectedMerchantIntegration: DemoSettings.merchantIntegration,
-            paymentType: .paypal
+            paymentType: .paypal,
+            appSwitchURL: Environment.sandbox.baseURL
         )
 
         state.paypalVaultTokenResponse = .loading