+++ Sources/CorePayments/Networking/AuthenticationTokenResponse.swift
@@ -3,6 +3,6 @@ import Foundation
 @_documentation(visibility: private)
 public struct AuthTokenResponse: Decodable {
 
-    let accessToken: String
+    public let accessToken: String
     let tokenType: String
 }